rename component image
