+++ src/components/get-body.js
@@ -1,0 +1,6 @@
+// Create body component
+export default function () {
+  // if (footContent == "" || !footContent.length) {
+  //   throw new Error("No footContent footer");
+  // }
+}
